+++ tests/item_tests/item_combat_coverage_test.rs
@@ -664,8 +664,8 @@ fn generate_subzone_boss_has_higher_stats() {
     for _ in 0..samples {
         let boss = generate_subzone_boss(zone, subzone);
         let mob = generate_zone_enemy(zone, subzone);
-        boss_hp_sum += boss.max_hp as u64;
-        mob_hp_sum += mob.max_hp as u64;
+        boss_hp_sum += boss.max_hp;
+        mob_hp_sum += mob.max_hp;
     }
     assert!(
         boss_hp_sum > mob_hp_sum,
@@ -711,8 +711,8 @@ fn generate_dungeon_elite_stronger_than_normal() {
     let mut normal_hp_sum = 0u64;
     let samples = 20;
     for _ in 0..samples {
-        elite_hp_sum += generate_dungeon_elite(5).max_hp as u64;
-        normal_hp_sum += generate_dungeon_enemy(5).max_hp as u64;
+        elite_hp_sum += generate_dungeon_elite(5).max_hp;
+        normal_hp_sum += generate_dungeon_enemy(5).max_hp;
     }
     assert!(
         elite_hp_sum > normal_hp_sum,
@@ -728,8 +728,8 @@ fn generate_dungeon_boss_stronger_than_elite() {
     let mut elite_hp_sum = 0u64;
     let samples = 20;
     for _ in 0..samples {
-        boss_hp_sum += generate_dungeon_boss(5).max_hp as u64;
-        elite_hp_sum += generate_dungeon_elite(5).max_hp as u64;
+        boss_hp_sum += generate_dungeon_boss(5).max_hp;
+        elite_hp_sum += generate_dungeon_elite(5).max_hp;
     }
     assert!(
         boss_hp_sum > elite_hp_sum,
@@ -778,8 +778,8 @@ fn zone_enemies_scale_with_zone() {
     let mut z5_hp = 0u64;
     let samples = 20;
     for _ in 0..samples {
-        z1_hp += generate_zone_enemy(zone1, &zone1.subzones[0]).max_hp as u64;
-        z5_hp += generate_zone_enemy(zone5, &zone5.subzones[0]).max_hp as u64;
+        z1_hp += generate_zone_enemy(zone1, &zone1.subzones[0]).max_hp;
+        z5_hp += generate_zone_enemy(zone5, &zone5.subzones[0]).max_hp;
     }
     assert!(
         z5_hp > z1_hp,
